udate2

+++ loadDatabase/F_QLSinhVien.cs
@@ -26,7 +26,7 @@ namespace loadDatabase
         public F_QLSinhVien()
         {
             InitializeComponent();
-           
+
 
             connection = new DB();
             primaryKey = new DataColumn[1];
@@ -50,7 +50,7 @@ namespace loadDatabase
             txtName.Text = DGVLSv.Rows[i].Cells[1].Value.ToString();
 
             txtClass.Text = DGVLSv.Rows[i].Cells[3].Value.ToString();
-            
+
         }
 
         void LoadDGVSinhVien()
@@ -60,9 +60,9 @@ namespace loadDatabase
                 var table = connection.GetDataTable(ma, "SinhVien");
                 DGVLSv.DataSource = table;
             }
-            catch(Exception ex) {
+            catch (Exception ex) {
                 MessageBox.Show("Lỗi");
-            } 
+            }
         }
 
         private void btnTim_Click(object sender, EventArgs e)
@@ -70,7 +70,7 @@ namespace loadDatabase
             TimSinhVien();
         }
 
-        public void LoadCbxLop() 
+        public void LoadCbxLop()
         {
             try
             {
@@ -92,63 +92,63 @@ namespace loadDatabase
         {
             try
             {
-               
 
-                    if (string.IsNullOrEmpty(txtName.Text))
+
+                if (string.IsNullOrEmpty(txtName.Text))
+                {
+                    MessageBox.Show("Vui Lòng Nhập Tên Sinh Viên ");
+                }
+                else
+                {
+                    if (string.IsNullOrEmpty(cbxLop.Text))
                     {
-                        MessageBox.Show("Vui Lòng Nhập Tên Sinh Viên ");
+                        MessageBox.Show("Vui Lòng Chọn Lớp !", "Thông Báo");
                     }
                     else
                     {
-                        if (string.IsNullOrEmpty(cbxLop.Text))
+                        Random rand = new Random();
+                        string kytu = "SV";
+                        string s = "";
+                        int temp = 9;
+                        string number;
+                        int num = 9;
+                        number = num.ToString();
+                        List<int> list = new List<int>();
+                        for (int i = 0; i <= 8; i++)
                         {
-                            MessageBox.Show("Vui Lòng Chọn Lớp !", "Thông Báo");
+                            list.Add(i);
+                            //list.Add(rand.Next(a));
+                        }
+                        //random
+                        for (int i = 0; i <= 8; i++)
+                        {
+                            temp = rand.Next(list.Count);
+                            s += list[temp];
+                            list.RemoveAt(temp);
+
+                        }
+                        string MSSV = kytu + s;
+                        LoadDGVSinhVien();
+                        DialogResult dlr = MessageBox.Show("MSSV :" + MSSV + "\nTên Sinh Viên: " + txtName.Text + "\n Mã Lơp :" + cbxLop.Text + "", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                        if (dlr == DialogResult.Yes)
+                        {
+                            string MuaVe = "insert into SinhVien values('" + MSSV + "',N'" + txtName.Text + "','2002/08/13','" + cbxLop.Text + "')";
+                            var table = connection.GetDataTable(MuaVe, "Ve");
+                            DGVLSv.AllowUserToAddRows = true;
+                            DGVLSv.DataSource = table;
+                            LoadDGVSinhVien();
                         }
                         else
                         {
-                            Random rand = new Random();
-                            string kytu = "SV";
-                            string s = "";
-                            int temp = 9;
-                            string number;
-                            int num = 9;
-                            number = num.ToString();
-                            List<int> list = new List<int>();
-                            for (int i = 0; i <= 8; i++)
-                            {
-                                list.Add(i);
-                                //list.Add(rand.Next(a));
-                            }
-                            //random
-                            for (int i = 0; i <= 8; i++)
-                            {
-                                temp = rand.Next(list.Count);
-                                s += list[temp];
-                                list.RemoveAt(temp);
 
-                            }
-                            string MSSV = kytu + s;
-                        LoadDGVSinhVien();
-                            DialogResult dlr = MessageBox.Show("MSSV :" + MSSV + "\nTên Sinh Viên: " + txtName.Text + "\n Mã Lơp :" + cbxLop.Text +"","",MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-                            if (dlr == DialogResult.Yes)
-                            {
-                            string MuaVe = "insert into SinhVien values('"+ MSSV + "',N'"+txtName.Text+"','2002/08/13','"+ cbxLop.Text+ "')";
-                                var table = connection.GetDataTable(MuaVe, "Ve");
-                                DGVLSv.AllowUserToAddRows = true;
-                                DGVLSv.DataSource = table;
-                                LoadDGVSinhVien();
-                            }
-                            else
-                            {
-
-                            }
                         }
                     }
-                
+                }
+
             }
-            catch (Exception ex) 
-            { 
-            
+            catch (Exception ex)
+            {
+
             }
 
         }
@@ -157,7 +157,28 @@ namespace loadDatabase
         {
             try
             {
-                string HVe = "Delete from SinhVien where Masv like'%" + txtMSSV.Text + "%'";
+
+                DialogResult dlr = MessageBox.Show("MSSV :" + txtMSSV.Text + "\nTên Sinh Viên: " + txtName.Text + "\n Mã Lơp :" + cbxLop.Text + "Bạn Muốn Xóa", "Xác Nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (dlr == DialogResult.Yes)
+                {
+                    string HVe = "Delete from SinhVien where Masv like'%" + txtMSSV.Text + "%'";
+                    var table = connection.GetDataTable(HVe, "SinhVien");
+                    DGVLSv.DataSource = table;
+                    LoadDGVSinhVien();
+                }
+              
+            }
+            catch (Exception ex)
+            {
+
+            }
+
+        }
+        public void UpdateSinhVien()
+        {
+            try
+            {
+                string HVe = "update SinhVien set HoTen =N'"+txtName.Text+"' where Masv='"+txtMSSV.Text+"'";
                 var table = connection.GetDataTable(HVe, "SinhVien");
                 DGVLSv.DataSource = table;
                 LoadDGVSinhVien();
@@ -168,7 +189,10 @@ namespace loadDatabase
 
             }
 
-        }
+
+        }  
+
+
 
         private void btnAdd_Click(object sender, EventArgs e)
         {
@@ -178,6 +202,11 @@ namespace loadDatabase
         private void btnXoa_Click(object sender, EventArgs e)
         {
             XoaSinhVien();
+        }
+
+        private void btnUpDate_Click(object sender, EventArgs e)
+        {
+            UpdateSinhVien();
         }
 
         void TimSinhVien()
